+++ static/src/components/FHIRECGAnnotationTaskView.vue
@@ -1,0 +1,22 @@
+<script lang="ts">
+import { defineComponent } from "vue";
+
+export default defineComponent({
+    data() {
+        return {
+            observationId: undefined as string | void,
+        };
+    },
+
+    methods:{
+        async nextObservation() {
+            // get next observation stuff
+        },
+    }
+})
+</script>
+
+
+<template>
+    // iFrame stuff
+</template>
