+++ static/src/components/TaskView.vue
@@ -1,20 +1,21 @@
 <script lang="ts">
 import { defineComponent } from "vue";
-import ImageDisplay from "@/components/ImageDisplay.vue";
-import AnnotationButton from "@/components/AnnotationButton.vue";
 import ProofOfCondition from "@/components/ProofOfCondition.vue";
 import UserInformationLabel from "@/components/UserInformationLabel.vue";
+import ImageClassificationTaskView from "@/components/ImageClassificationTaskView.vue";
+import FHIRECGAnnotationTaskView from "@/components/FHIRECGAnnotationTaskView.vue";
 import { TasksService as API } from "@/api/services/TasksService";
-import type { Task } from "@/api/models/Task";
+import type { BaseTask } from "@/api/models/BaseTask";
+import type { ImageTask } from "@/api/models/ImageTask";
 import { paramCase } from "change-case";
 import { fullname } from "@/utils/helpers";
 
 export default defineComponent({
   components: {
-    ImageDisplay,
-    AnnotationButton,
     ProofOfCondition,
     UserInformationLabel,
+    ImageClassificationTaskView,
+    FHIRECGAnnotationTaskView,
   },
 
   data() {
@@ -24,8 +25,7 @@ export default defineComponent({
       isTrained: false,
       isLoading: false,
       isError: false,
-      imageId: undefined as string | void,
-      task: undefined as Task | void,
+      task: undefined as BaseTask | void,
     };
   },
 
@@ -41,20 +41,7 @@ export default defineComponent({
         const id = this.$route.params["taskId"];
         if (typeof id !== "string") throw Error();
         this.task = await API.getTask(id);
-        await this.nextImage();
-      } catch {
-        this.isError = true;
-      } finally {
-        this.isLoading = false;
-      }
-    },
-
-    async nextImage() {
-      try {
-        this.isLoading = true;
-        this.imageId = undefined;
-        if (!this.task) throw Error();
-        this.imageId = await API.getNextImage(this.task.id).catch(console.log);
+        console.log(this.task);
       } catch {
         this.isError = true;
       } finally {
@@ -82,33 +69,11 @@ export default defineComponent({
     <div v-else-if="isLoading">
       <i-loader />
     </div>
-    <i-card
-      v-else-if="imageId === undefined"
-      id="no-more-images"
-      class="margin-y:20px"
-    >
-      No more Images to label
-    </i-card>
-    <div v-else>
-      <ImageDisplay
-        id="image-display"
-        :task-id="task!.id"
-        :image-id="imageId"
-      />
-      <i-button-group block>
-        <AnnotationButton
-          v-for="label in task!.labels"
-          :id="'annotation-button-' + paramCase(label)"
-          :key="label"
-          :label="label"
-          :is-attentive="isAttentive"
-          :is-trained="isTrained"
-          :competency="'Prof. Dr. Med.'"
-          :task-id="task!.id"
-          :image-id="imageId!"
-          @annotation-saved="nextImage"
-        />
-      </i-button-group>
-    </div>
+    <ImageClassificationTaskView
+      v-else
+      :isAttentive="isAttentive"
+      :isTrained="isTrained"
+      :task="(task as any)"
+    />
   </div>
 </template>
